depth_upscaling/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy
scipy
matplotlib
webdataset
open3d
timm
einops
Pillow

